Initial Commit
The commit includes the basic setup for a Spring Boot project. It
includes the gitignore file, the build.gradle file containing all
the dependencies, as well as the AppRunner class, which will allow
the application to run. The gradle wrapper has been added as well for
ease of deployment/build.

The application is set to run on port 6001.