Replace RequestMatcher approach to check Authority in SecurityConfig
Previously, in SecurityConfig, the RequestMatchers were checking for the role
using the `hasRole` method, which requires for the roles to have the "ROLE_" prefix.
However, the application is not designed for the roles to have the said prefix.

As such, the check has been replaced with a more appropriate `hasAuthority` method.