Introduced LoginBaseDto and Refactored RegistrationValidator
As a part of introducing user registration from the Admin Dashboard, the RegistrationRequestDto
had to be refactored. A new base DTO called LoginBaseDto has been added, which
captures all the basic information of the login. RegistrationRequestDto now contains
only Confirm Password field, which is specific to it.

Also, the RegistrationValidator has been renamed to LoginDtoValidator to be more
generic - i.e. be able to validate any DTO that extends LoginBaseDto.