Implement Admin :: Genre Form & List
This involves creating GenreAdminController, which houses the handler methods
to navigate to the Genre List and the Genre Form.

The list shows the name of the genres added, along with the date/time when they
were added. The list also includes an edit button, which takes admin users to
Genre form.

The genre form is used by both the new and edit endpoints, allowing users to create
and modify Genres respectively.