Implement Book List page
The Book List page has been implemented to show all the books stored in the application.
For starters, the list is sorted by the book's title, and at a time, 18 books are shown
in the page. The BookListRepository implemented previously allows for pagination in the
page.

The "Browse" link the navbar has been updated to show all the books stored in the
application.